Add total bandwidth report for single threaded tests
The single threaded tests multiplex the time over all the channels.
Previously only the average bandwidth for each channel was reported.
Including the aggregated bandwidth of all channels in the report not
only provides a metric that matches more closely with the transport
capacity, but also is more consistent with the result from the multi-
threaded test.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>